// reactapp/src/services/api.ts
const BASE_PATH = "/".replace(/\/+$/, "");

/**
 *
 * @export
 */
export const COLLECTION_FORMATS = {
    csv: ",",
    ssv: " ",
    tsv: "\t",
    pipes: "|",
};

/**
 *
 * @export
 * @interface FetchAPI
 */
export interface FetchAPI {
    (url: string, init?: any): Promise<Response>;
}

/**
 *
 * @export
 * @interface FetchArgs
 */
export interface FetchArgs {
    url: string;
    options: any;
}

/**
 *
 * @export
 * @class BaseAPI
 */
export class BaseAPI {
    protected configuration: Configuration | undefined;

    constructor(configuration?: Configuration, protected basePath: string = BASE_PATH, protected fetch: FetchAPI = isomorphicFetch) {
        if (configuration) {
            this.configuration = configuration;
            this.basePath = configuration.basePath || this.basePath;
        }
    }
};

/**
 *
 * @export
 * @class RequiredError
 * @extends {Error}
 */
export class RequiredError extends Error {
    name: "RequiredError" = "RequiredError"
    constructor(public field: string, msg?: string) {
        super(msg);
    }
}

/**
 * 
 * @export
 * @interface Hello
 */
export interface Hello {
    /**
     * 
     * @type {number}
     * @memberof Hello
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Hello
     */
    name?: string;
}
/**
 * 
 * @export
 * @interface WeatherForecast
 */
export interface WeatherForecast {
    /**
     * 
     * @type {Date}
     * @memberof WeatherForecast
     */
    date?: Date;
    /**
     * 
     * @type {number}
     * @memberof WeatherForecast
     */
    temperatureC?: number;
    /**
     * 
     * @type {number}
     * @memberof WeatherForecast
     */
    temperatureF?: number;
    /**
     * 
     * @type {string}
     * @memberof WeatherForecast
     */
    summary?: string;
}
// ...
export class HelloApi extends BaseAPI {
    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof HelloApi
     */
    public apiHelloGet(options?: any) {
        return HelloApiFp(this.configuration).apiHelloGet(options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof HelloApi
     */
    public apiHelloIdDelete(id: number, options?: any) {
        return HelloApiFp(this.configuration).apiHelloIdDelete(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof HelloApi
     */
    public apiHelloIdGet(id: number, options?: any) {
        return HelloApiFp(this.configuration).apiHelloIdGet(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {number} id 
     * @param {Hello} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof HelloApi
     */
    public apiHelloIdPut(id: number, body?: Hello, options?: any) {
        return HelloApiFp(this.configuration).apiHelloIdPut(id, body, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {Hello} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof HelloApi
     */
    public apiHelloPost(body?: Hello, options?: any) {
        return HelloApiFp(this.configuration).apiHelloPost(body, options)(this.fetch, this.basePath);
    }

}
// ...
export class WeatherForecastApi extends BaseAPI {
    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof WeatherForecastApi
     */
    public getWeatherForecast(options?: any) {
        return WeatherForecastApiFp(this.configuration).getWeatherForecast(options)(this.fetch, this.basePath);
    }

}